+++ ui/src/common/ArtistSimpleList.js
