+++ ui/src/common/index.js
@@ -27,4 +27,4 @@ export * from './useToggleStar'
 export * from './useTraceUpdate'
 export * from './Writable'
 export * from './SongSimpleList'
-export * from './artistSimpleList'
+export * from './ArtistSimpleList'
